+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Consoles
 - Update `ricaun.Autodesk.Forge.Oss.DesignAutomation` to `3.1.0-*`.
 - Add `AccessToken` to pass the `adsk3LeggedToken` in DA4R.
+- Add `AspUserInfoUtils` to get Autodesk user info using `AccessToken`.
 - Remove local environment variables.
 ### Tests
 - Add `UserIdTests` to test `adsk3LeggedToken` in DA4R.